Add some return statements
These return statements are not really needed here,
but lets just unify the array handling with that of
the strings.

Advantage is that if code is executed after the if
blocks, we could find them easily by adding a line
that throws an error (if we want) or by doing code
coverage.

Another advantage is that we could add more if
statements after array handling, in case we have to
deal with more types.

So, this patch just adds 2 silly lines of code that may
be seen as "just stopid", but just add it, as its not
changing the outside world behavior right?